Atualizando lists properties

+++ statistics/lists_alters_overlap_v1.0.py
@@ -65,21 +65,21 @@ def get_alters_set(ego_id,file):
 			if len(a) > 2:
 				if long(a[0])==long(ego_id):
 					alters_set.add(long(a[0]))										# Pega só o primeiro campo - nó1
-					tuple=(a[1],a[2])													# Armazena o peso do relacionamento entre o alter e o ego para formar um ranking depois.
+					tuple=(long(a[1]),float(a[2]))													# Armazena o peso do relacionamento entre o alter e o ego para formar um ranking depois.
 					ranking.append(tuple)
 				elif long(a[1])==long(ego_id):
 					alters_set.add(long(a[1]))										# Pega só o segundo campo - nó 2
-					tuple=(a[0],a[2])													# Armazena o peso do relacionamento entre o alter e o ego para formar um ranking depois.
+					tuple=(long(a[0]),float(a[2]))													# Armazena o peso do relacionamento entre o alter e o ego para formar um ranking depois.
 					ranking.append(tuple)
 			else:																			# Rede sem peso
 				weight = 1
 				if long(a[0])==long(ego_id):
 					alters_set.add(long(a[0]))										# Pega só o primeiro campo - nó1
-					tuple=(a[1],weight)												# Armazena o peso do relacionamento entre o alter e o ego para formar um ranking depois.
+					tuple=(long(a[1]),float(weight))												# Armazena o peso do relacionamento entre o alter e o ego para formar um ranking depois.
 					ranking.append(tuple)
 				elif long(a[1])==long(ego_id):
 					alters_set.add(long(a[1]))										# Pega só o segundo campo - nó 2
-					tuple=(a[0],weight)												# Armazena o peso do relacionamento entre o alter e o ego para formar um ranking depois.
+					tuple=(long(a[0]),float(weight))												# Armazena o peso do relacionamento entre o alter e o ego para formar um ranking depois.
 					ranking.append(tuple)				
 				
 	if ranking is not None:
